* Added fileAge support - that is, miniumum age (ms) for a file to be processed.  Changed the implementation of 'hasChanged' from doing 2 consecutive calls to get the file size to using lastModifiedTime together with fileAge * Added temp directory support -  The file is first uploaded to a temporary directory. When the file is fully uploaded the file is moved to its final destination. The tempDir will be created as a sub directory to the endpoint. * Fixed so that the file in the inbound directory is not deleted if some error occurs when writing to the outbound directory * Added support for absolute paths - sftp:user@host//opt/foo/bar mean /opt/foo/bar.

+++ src/main/java/org/mule/transport/sftp/SftpInputStream.java
@@ -1,3 +1,13 @@
+/*
+ * $Id$
+ * --------------------------------------------------------------------------------------
+ * Copyright (c) MuleSource, Inc.  All rights reserved.  http://www.mulesource.com
+ *
+ * The software in this package is published under the terms of the MuleSource MPL
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+
 package org.mule.transport.sftp;
 
 import java.io.IOException;
@@ -15,6 +25,7 @@ public class SftpInputStream extends InputStream
 	SftpClient client;
 	boolean autoDelete = true;
 	String fileName;
+	boolean errorOccured = false;
 
 	/**
 	 * A special sftp InputStream.  The constuctor creates the InputStream by
@@ -62,7 +73,8 @@ public class SftpInputStream extends InputStream
 
 		is.close();
 
-		if( autoDelete )
+
+		if( autoDelete && !errorOccured)
 		{
 		    client.deleteFile(fileName);
 		}
@@ -72,7 +84,8 @@ public class SftpInputStream extends InputStream
 	}
 
 
-
-
-
+	public void setErrorOccured(boolean errorOccured)
+	{
+		this.errorOccured = errorOccured;
+	}
 }
